+++ src/tests/ExtraLINQ.Tests/IEnumerable/HasAtMostTests.cs
@@ -6,14 +6,14 @@ using Xunit;
 
 namespace ExtraLinq.Tests
 {
-    public class CountsMaxTests
+    public class HasAtMostTests
     {
         [Fact]
         public void ThrowsArgumentNullExceptionWhenCollectionIsNull()
         {
             IEnumerable<object> nullCollection = null;
 
-            Assert.Throws<ArgumentNullException>(() => nullCollection.CountsMax(1));
+            Assert.Throws<ArgumentNullException>(() => nullCollection.HasAtMost(1));
         }
 
         [Fact]
@@ -21,7 +21,7 @@ namespace ExtraLinq.Tests
         {
             IEnumerable<char> letters = "abcde";
 
-            Assert.Throws<ArgumentOutOfRangeException>(() => letters.CountsMax(-10));
+            Assert.Throws<ArgumentOutOfRangeException>(() => letters.HasAtMost(-10));
         }
 
         [Fact]
@@ -30,14 +30,14 @@ namespace ExtraLinq.Tests
             IEnumerable<char> letters = "abcd";
             IEnumerable<char> emptyCollection = Enumerable.Empty<char>();
 
-            letters.CountsMax(4).Should().BeTrue();
-            letters.CountsMax(5).Should().BeTrue();
-            emptyCollection.CountsMax(0).Should().BeTrue();
+            letters.HasAtMost(4).Should().BeTrue();
+            letters.HasAtMost(5).Should().BeTrue();
+            emptyCollection.HasAtMost(0).Should().BeTrue();
 
             // Test ICollection.Count early exit strategy
-            letters.ToList().CountsMax(4).Should().BeTrue();
-            letters.ToList().CountsMax(5).Should().BeTrue();
-            emptyCollection.ToList().CountsMax(0).Should().BeTrue();
+            letters.ToList().HasAtMost(4).Should().BeTrue();
+            letters.ToList().HasAtMost(5).Should().BeTrue();
+            emptyCollection.ToList().HasAtMost(0).Should().BeTrue();
         }
 
         [Fact]
@@ -45,12 +45,12 @@ namespace ExtraLinq.Tests
         {
             IEnumerable<char> letters = "abcd";
 
-            letters.CountsMax(4).Should().BeTrue();
-            letters.CountsMax(5).Should().BeTrue();
+            letters.HasAtMost(4).Should().BeTrue();
+            letters.HasAtMost(5).Should().BeTrue();
 
             // Test ICollection.Count early exit strategy
-            letters.ToList().CountsMax(4).Should().BeTrue();
-            letters.ToList().CountsMax(5).Should().BeTrue();
+            letters.ToList().HasAtMost(4).Should().BeTrue();
+            letters.ToList().HasAtMost(5).Should().BeTrue();
         }
 
         [Fact]
@@ -59,7 +59,7 @@ namespace ExtraLinq.Tests
             IEnumerable<object> nullCollection = null;
             Func<object, bool> alwaysTruePredicate = _ => true;
 
-            Assert.Throws<ArgumentNullException>(() => nullCollection.CountsMax(1, alwaysTruePredicate));
+            Assert.Throws<ArgumentNullException>(() => nullCollection.HasAtMost(1, alwaysTruePredicate));
         }
 
         [Fact]
@@ -67,7 +67,7 @@ namespace ExtraLinq.Tests
         {
             IEnumerable<char> validCollection = "abcd";
 
-            Assert.Throws<ArgumentNullException>(() => validCollection.CountsMax(1, null));
+            Assert.Throws<ArgumentNullException>(() => validCollection.HasAtMost(1, null));
         }
 
         [Fact]
@@ -76,7 +76,7 @@ namespace ExtraLinq.Tests
             IEnumerable<char> validCollection = "abcd";
             Func<char, bool> validPredicate = c => c == 'a';
 
-            Assert.Throws<ArgumentOutOfRangeException>(() => validCollection.CountsMax(-1, validPredicate));
+            Assert.Throws<ArgumentOutOfRangeException>(() => validCollection.HasAtMost(-1, validPredicate));
         }
 
         [Fact]
@@ -85,9 +85,9 @@ namespace ExtraLinq.Tests
             IEnumerable<string> fruits = new[] { "apple", "apricot", "banana" };
             Func<string, bool> startsWithLowercasedA = fruit => fruit.StartsWith("a");
 
-            fruits.CountsMax(2, startsWithLowercasedA).Should().BeTrue();
-            fruits.CountsMax(3, startsWithLowercasedA).Should().BeTrue();
-            fruits.CountsMax(int.MaxValue, startsWithLowercasedA).Should().BeTrue();
+            fruits.HasAtMost(2, startsWithLowercasedA).Should().BeTrue();
+            fruits.HasAtMost(3, startsWithLowercasedA).Should().BeTrue();
+            fruits.HasAtMost(int.MaxValue, startsWithLowercasedA).Should().BeTrue();
         }
 
         [Fact]
@@ -96,8 +96,8 @@ namespace ExtraLinq.Tests
             IEnumerable<string> fruits = new[] { "apple", "apricot", "banana" };
             Func<string, bool> startsWithLowercasedA = fruit => fruit.StartsWith("a");
 
-            fruits.CountsMax(0, startsWithLowercasedA).Should().BeFalse();
-            fruits.CountsMax(1, startsWithLowercasedA).Should().BeFalse();
+            fruits.HasAtMost(0, startsWithLowercasedA).Should().BeFalse();
+            fruits.HasAtMost(1, startsWithLowercasedA).Should().BeFalse();
         }
     }
 }