+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #### 2.0.0-alpha1 - ??
 * Removed `ElementAt` method accepting an `IndexingStrategy` (use `Cycle` instead)
 * Renamed `CountsExactly` to `HasExactly`
+* Renamed `CountsMax` to `HasAtMost`
 * Added `Chunk` method
 * Added `Cycle` method
 * Added `Distinct` method